Enable the default panic unwind hook to print backtraces again
The default panic unwind hook attempts to print the backtrace by
first initializing DbgHelp with `SymInitializeW`. The documentation
states that this can't be called again until after `SymCleanup` is
called first. If called prior to cleanup, it returns a non-success
result. The default unwind hook detects this error and never tries
to print the backtrace.

This PR installs a new panic handler that first calls `SymCleanup`
prior to calling the previously installed (or default) panic hook.

Fixes #165