Switch to the `object` crate for object parsing
This commit switches the gimli feature from the `goblin` crate to the
`object` crate for parsing object files. The main motivation here is
trimming the dependencies of the `gimli-symbolize` feature to a bare
minimum. The `object` crate itself has no dependencies now and should be
a relatively easy drop-in replacement for the `goblin` crate.