Fix macOS symbolication of system libraries
This commit fixes an issue where symbolication of system libraries
didn't work on macOS. Symbolication through the symbol table was always
off by a slide amount for the library. It's not entirely clear why this
kept happening or what was going on, but some poking in LLDB's source
revealed a way we can differentiate and figure out what addresses need
to be looked up in the symbol table. Some more information is contained
in the comments of the commit itself.

Closes #318