Remove the `unix-backtrace` feature
AFAIK there's not actual use case for this. This was added ages ago when
backtrace functionality was first implemented but we've always preferred
libunwind as an unwinder since it's more accurate. No platform uses this
by default today, so this commit is removing the support and we can
re-add it pretty easily if necessary for a new platform.