Make libunwind/dbghelp features defunkt
AFAIK there's no current users of this crate who are tweaking the
various methods of capturing a backtrace or symbolizing a backtrace. I
believe everyone's just largely using the default features. Additionally
it would be extremely hard to *not* use the default features given the
wide usage of this crate.

Given all this this commit simply removes the option of doing so.
Instead this crate will automatically select the best implementation
strategy per-platform and will provide the necessary support to enable
it. It's intended to that this should give a much better default
experience the avoids futzing around with features so much.