sys: Split dep-of-std and symbol prefix features
When building a libstd port out-of-tree using xargo's staged build
system, libstd dependencies are allowed to rely on the presence of
libcore, compiler-builtins, etc... in the sysroot. Such builds will thus
want to remove the rustc-dep-of-std feature. Doing so for backtrace-sys,
however, risks causing duplicate symbols between the sysroot and
normal libraries. This is because backtrace-sys links in the
libbacktrace library, and has some special logic associated with the
rustc-dep-of-std feature to rename the symbols.

In order to better support such libstd crates, this commit splits the
symbol renaming logic into its own feature, and makes that feature a
dependency of rustc-dep-of-std.