Fix get_sp implementation for s390x
On s390x we use a biased CFA: the CFA of a function equal the SP
at function entry plus a constant of 160.

Therefore, the current implementation of get_sp, which uses
_Unwind_GetCFA, is not an appropriate way to retrieve the
current SP value on this platform.

Instead, use _Unwind_GetGR to explicitly retrieve the value
of the stack pointer register %r15.