Added tracing of threads other than the calling thread
This probably wont compile on anything other than windows currently.

Previously RtlCaptureContext was used to effortlessly get the calling thread's context.
Instead pass in a handle to trace and then use GetThreadContext to get the context of the thread.

This should allow reading the backtrace of threads other than the calling thread. This is needed in the windows implementation of pprof-rs.

Some notes:
- The thread handle must be opened with THREAD_GET_CONTEXT access (see windows docs)
- Thread must be suspended before the trace call was made (see windows docs)
- A crossplatform interface/implementation might not be possible.
- I have no idea whether the stacktrace is even valid. It makes sense if it is?