Allow the library consumer to specify the Android API level
The build.rs script will use the C preprocessor to pull the API level
out of C headers by default. But for build systems that wish to control
each build step, and rely on remote compilation, this presents problems.

Remove the requirement to run the C preprocessor if the API level is
specified via `--cfg`:
* android_api_at_least_0: No minimum API level is guaranteed.
* android_api_at_least_21: The API level will be at least 21.