Support zstd-compressed ELF sections.
zstd has been introduced as an alternative to zlib for the compression of debug
sections.[0] Toolchain support is widely present at this time but lack of
support in backtrace is a severe limitation on using this feature in Rust
programs.

Unlike zlib compression, the most widely used Rust library for zstd is a set
of bindings to the zstd C library. There does exist a Rust reimplementation
(ruzstd) however the author claims it is significantly slower than the C
library and it sees substantially less usage, so I have not used it here.

Using the C library poses challenges because it requires compiling native
code. I made a few adjustments to the CI configuration to make that possible
on all platforms, and pinned the zstd crate to a slightly older version
due to gyscos/zstd-rs#271 breaking the crate on wasm targets on the very
latest version. Additionally, I've disabled this entirely on illumos rather
than try to get native code to compile successfully there.

[0] https://maskray.me/blog/2022-09-09-zstd-compressed-debug-sections