ci: remove binary size check
Remove the "binary size check" workflow because it is often broken and
now and then makes me sit through a discussion about its security needs.
These discussions always make me feel very silly.

I originally hoped this CI job would be a useful part of assuring that
we would not accidentally bloat binaries when backtrace merges into std.
However, it feels like that is not achieved by this workflow and at some
point it has definitely outweighed the cost of maintaining it.

We can readd something if we find a better way to implement things which
is notably less jank and gets less complaints.