Significantly reduce the size of Cache::with_global
Previously this would copy the freshly created cache into the static.
Given that Cache is fairly big and for some reason LLVM decided to
unroll the memcpy despite optimize(size), this produced a lot of code.
This commit however changes it to create a partially initialized Cache
at compile time and only initialize a single Vec field when accessing it
for the first time.

This reduces the size of libas_if_std.rlib from 205kb to 199kb.